Write additional tests.
