+++ src/app/tabs/home/containers/SearchBoxContainer.test.js
@@ -3,18 +3,12 @@ import { mount } from 'enzyme';
 import { Provider } from 'react-redux';
 
 import searchReducer from '../../search/reducer';
-import searchOperation from '../../search/operations';
-
 import SearchBoxContainer from './SearchBoxContainer';
 
 import multiLanguageStore from '../../../../../tests/config/multiLanguageStore';
 
 describe('searchBoxContainer', () => {
   const storeSetup = { search: searchReducer };
-
-  it('gets correct environment varialbe', () => {
-    expect(process.env.REACT_APP_ENVIRONMENT).toEqual('test');
-  });
 
   it('has default state', () => {
     const store = multiLanguageStore(storeSetup);
@@ -35,19 +29,5 @@ describe('searchBoxContainer', () => {
 
     component.find('button').simulate('click');
     expect(store.getState().search.domain).toEqual('foobar');
-  });
-
-  it('searches for the domain via handleClick opperations', () => {
-    const store = multiLanguageStore(storeSetup);
-    console.log('setting up man!');
-    return store.dispatch(searchOperation('jesse'))
-      .then(() => {
-        const searchState = store.getState().search;
-
-        expect(searchState.domain).toEqual('jesse');
-        expect(searchState.requestingOwner).toBeFalsy();
-        expect(searchState.owner).toBe('asd');
-        expect(searchState.owned).toBeTruthy();
-      });
   });
 });
