+++ tests/setEnvVars.test.js
@@ -1,0 +1,5 @@
+describe('setEnvVars', () => {
+  it('gets correct environment varialbe', () => {
+    expect(process.env.REACT_APP_ENVIRONMENT).toEqual('test');
+  });
+});
